+++ Sora/PeerChannel.swift
@@ -110,6 +110,7 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
 
     var webRTCConfiguration: WebRTCConfiguration
     var clientId: String?
+    var bundleId: String?
     var connectionId: String?
 
     var onConnectHandler: ((Error?) -> Void)?
@@ -288,11 +289,12 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
 
         let webRTCVersion = "Shiguredo-build \(WebRTCInfo.version) (\(WebRTCInfo.version).\(WebRTCInfo.commitPosition).\(WebRTCInfo.maintenanceVersion) \(WebRTCInfo.shortRevision))"
 
-        let simulcast = configuration.simulcastEnabled || configuration.spotlightEnabled == .enabled
+        let simulcast = configuration.simulcastEnabled
         let connect = SignalingConnect(
             role: role,
             channelId: configuration.channelId,
             clientId: configuration.clientId,
+            bundleId: configuration.bundleId,
             metadata: configuration.signalingConnectMetadata,
             notifyMetadata: configuration.signalingConnectNotifyMetadata,
             sdp: sdp,
@@ -385,15 +387,11 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
                 }
             }
         } else {
-            // mid なし
-            if let track = stream.nativeVideoTrack {
-                nativeChannel.add(track,
-                                  streamIds: [stream.nativeStream.streamId])
-            }
-            if let track = stream.nativeAudioTrack {
-                nativeChannel.add(track,
-                                  streamIds: [stream.nativeStream.streamId])
-            }
+            // mid なしの場合はエラーにする
+            Logger.error(type: .peerChannel, message: "mid not found")
+            disconnect(error: SoraError.peerChannelError(reason: "mid not found"),
+                       reason: .signalingFailure)
+            return
         }
 
         // マイクの初期化
@@ -786,6 +784,7 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
             signalingChannel.setConnectedUrl()
 
             clientId = offer.clientId
+            bundleId = offer.bundleId
             connectionId = offer.connectionId
             if let dataChannels = offer.dataChannels {
                 signalingChannel.dataChannelSignaling = true