+++ Sora/Signaling.swift
@@ -11,24 +11,26 @@ private func serializeMetadataList(_ data: Any?) -> [SignalingNotifyMetadata]? {
 
     let result = array.map { (dict: [String: Any]) -> SignalingNotifyMetadata in
         var signalingNotifyMetadata = SignalingNotifyMetadata()
-        if dict.keys.contains("client_id"), let clinetId = dict["client_id"] as? String? {
+        if let clinetId = dict["client_id"] as? String {
             signalingNotifyMetadata.clientId = clinetId
         }
-
-        if dict.keys.contains("connection_id"), let connectionId = dict["connection_id"] as? String? {
+        if let bundleId = dict["bundle_id"] as? String {
+            signalingNotifyMetadata.bundleId = bundleId
+        }
+        if let connectionId = dict["connection_id"] as? String {
             signalingNotifyMetadata.connectionId = connectionId
         }
 
-        if dict.keys.contains("authn_metadata") {
-            signalingNotifyMetadata.authnMetadata = dict["authn_metadata"]
+        if let authnMetadata = dict["authn_metadata"] {
+            signalingNotifyMetadata.authnMetadata = authnMetadata
         }
 
-        if dict.keys.contains("authz_metadata") {
-            signalingNotifyMetadata.authzMetadata = dict["authz_metadata"]
+        if let authzMetada = dict["authz_metadata"] {
+            signalingNotifyMetadata.authzMetadata = authzMetada
         }
 
-        if dict.keys.contains("metadata") {
-            signalingNotifyMetadata.metadata = dict["metadata"]
+        if let metadata = dict["metadata"] {
+            signalingNotifyMetadata.metadata = metadata
         }
 
         return signalingNotifyMetadata
@@ -53,33 +55,33 @@ private func updateMetadata(signaling: Signaling, data: Data) -> Signaling {
     switch signaling {
     case var .offer(message):
         // TODO: if json.keys.contains("key") を if let に書き換えたい
-        if json.keys.contains("metadata") {
-            message.metadata = json["metadata"]
+        if let metadata = json["metadata"] {
+            message.metadata = metadata
         }
         if let dataChannels = json["data_channels"] as? [[String: Any]] {
             message.dataChannels = dataChannels
         }
         return .offer(message)
     case var .push(message):
-        if json.keys.contains("data") {
-            message.data = json["data"]
+        if let data = json["data"] {
+            message.data = data
         }
         return .push(message)
     case var .notify(message):
-        if json.keys.contains("authn_metadata") {
-            message.authnMetadata = json["authn_metadata"]
+        if let authnMetadata = json["authn_metadata"] {
+            message.authnMetadata = authnMetadata
         }
-        if json.keys.contains("authz_metadata") {
-            message.authzMetadata = json["authz_metadata"]
+        if let authzMetadata = json["authz_metadata"] {
+            message.authzMetadata = authzMetadata
         }
-        if json.keys.contains("metadata") {
-            message.metadata = json["metadata"]
+        if let metadata = json["metadata"] {
+            message.metadata = metadata
         }
-        if json.keys.contains("metadata_list") {
-            message.metadataList = serializeMetadataList(json["metadata_list"])
+        if let metadataList = json["metadata_list"] {
+            message.metadataList = serializeMetadataList(metadataList)
         }
-        if json.keys.contains("data") {
-            message.data = serializeMetadataList(json["data"])
+        if let data = json["data"] {
+            message.data = serializeMetadataList(data)
         }
         return .notify(message)
     default:
@@ -242,6 +244,9 @@ public struct SignalingNotifyMetadata {
     /// クライアント ID
     public var clientId: String?
 
+    /// バンドル ID
+    public var bundleId: String?
+
     /// 接続 ID
     public var connectionId: String?
 
@@ -268,6 +273,9 @@ public struct SignalingConnect {
 
     /// クライアント ID
     public var clientId: String?
+
+    /// バンドル ID
+    public var bundleId: String?
 
     /// メタデータ
     public var metadata: Encodable?
@@ -418,6 +426,9 @@ public struct SignalingOffer {
     /// クライアント ID
     public let clientId: String
 
+    /// バンドル ID
+    public let bundleId: String?
+
     /// 接続 ID
     public let connectionId: String
 
@@ -532,6 +543,9 @@ public struct SignalingNotify {
 
     /// クライアント ID
     public var clientId: String?
+
+    /// バンドル ID
+    public var bundleId: String?
 
     /// 接続 ID
     public var connectionId: String?
@@ -800,6 +814,7 @@ extension SignalingConnect: Codable {
         case role
         case channel_id
         case client_id
+        case bundle_id
         case metadata
         case signaling_notify_metadata
         case sdp
@@ -840,6 +855,7 @@ extension SignalingConnect: Codable {
         try container.encode(role, forKey: .role)
         try container.encode(channelId, forKey: .channel_id)
         try container.encodeIfPresent(clientId, forKey: .client_id)
+        try container.encodeIfPresent(bundleId, forKey: .bundle_id)
         try container.encodeIfPresent(sdp, forKey: .sdp)
         let metadataEnc = container.superEncoder(forKey: .metadata)
         try metadata?.encode(to: metadataEnc)
@@ -959,6 +975,7 @@ extension SignalingOffer.Encoding: Codable {
 extension SignalingOffer: Codable {
     enum CodingKeys: String, CodingKey {
         case client_id
+        case bundle_id
         case connection_id
         case sdp
         case config
@@ -969,6 +986,7 @@ extension SignalingOffer: Codable {
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         clientId = try container.decode(String.self, forKey: .client_id)
+        bundleId = try container.decodeIfPresent(String.self, forKey: .bundle_id)
         connectionId = try container.decode(String.self, forKey: .connection_id)
         sdp = try container.decode(String.self, forKey: .sdp)
         configuration =
@@ -1083,6 +1101,7 @@ extension SignalingNotify: Codable {
         case event_type
         case role
         case client_id
+        case bundle_id
         case connection_id
         case audio
         case video
@@ -1105,6 +1124,7 @@ extension SignalingNotify: Codable {
                                          forKey: .event_type)
         role = try container.decodeIfPresent(SignalingRole.self, forKey: .role)
         clientId = try container.decodeIfPresent(String.self, forKey: .client_id)
+        bundleId = try container.decodeIfPresent(String.self, forKey: .bundle_id)
         connectionId = try container.decodeIfPresent(String.self,
                                                      forKey: .connection_id)
         audioEnabled = try container.decodeIfPresent(Bool.self, forKey: .audio)